Build a performance table from MSE result files

Reads a set of serialized MSE result files from disk and combines their performance statistics into a single data.table.

getMSEPerformance(path, pattern = "*.rds")

Arguments

path

A character string giving the directory containing serialized MSE result files.

pattern

A character string giving the file pattern used by base::list.files() to identify files to read. Defaults to "*.rds".

Value

A data.table created by row-binding performance results from all matching files.

Details

For each file found in path matching pattern, the object is read using readRDS(). Then:

if the object already inherits from data.table, it is returned as-is;

otherwise, performance() is called on the object to derive the performance table.

All resulting tables are combined with data.table::rbindlist() using fill=TRUE, so files with slightly different column structures can still be merged.

This function is intended as a convenient batch importer for previously saved MSE outputs or already extracted performance tables.

Typical use cases

combining performance output from multiple management procedures;

importing a mixture of saved raw MSE objects and precomputed data.table performance summaries;

preparing inputs for writePerformance() or downstream summaries.

See also

Author

Iago Mosqueira, WMR

Examples

if (FALSE) { # \dontrun{
# Read and combine performance tables from a folder of MSE results
dat <- getMSEPerformance("mse")

# Use a custom file pattern
dat <- getMSEPerformance("mse", pattern = "scenario_.*\\.rds$")
} # }